Moved LookupTableItem class to separate java file
