+++ components/insight/SRC/org/openmicroscopy/shoola/util/ui/colourpicker/LUTUI.java
@@ -25,7 +25,10 @@ package org.openmicroscopy.shoola.util.ui.colourpicker;
 import java.awt.GridBagConstraints;
 import java.awt.GridBagLayout;
 import java.awt.Insets;
+import java.util.ArrayList;
+import java.util.Collections;
 import java.util.Iterator;
+import java.util.List;
 
 import javax.swing.JList;
 import javax.swing.JPanel;
@@ -36,8 +39,6 @@ import javax.swing.event.ListSelectionListener;
 
 import org.apache.commons.collections.CollectionUtils;
 
-import edu.emory.mathcs.backport.java.util.Arrays;
-
 /**
  * This is a UI which represents the lookup tables as a list.
  *
@@ -45,9 +46,6 @@ import edu.emory.mathcs.backport.java.util.Arrays;
  *         href="mailto:d.lindner@dundee.ac.uk">d.lindner@dundee.ac.uk</a>
  */
 public class LUTUI extends JPanel {
-
-    /** The name of the 'None' lookup table entry */
-    private static final String NONE = "None";
 
     /** List of lookup tables. */
     private JList lutList;
@@ -94,8 +92,10 @@ public class LUTUI extends JPanel {
         lutList.setSelectedIndex(index);
         selectionListener = new ListSelectionListener() {
             public void valueChanged(ListSelectionEvent e) {
-                LookupTableItem lut = (LookupTableItem) lutList.getSelectedValue();
-                if (lut == null || lut.equals(NONE))
+                LookupTableItem lut = (LookupTableItem) lutList
+                        .getSelectedValue();
+                if (lut == null || lut == LookupTableItem.NONE
+                        || lut == LookupTableItem.SEPARATOR)
                     c.setLUT(null);
                 else
                     c.setLUT(lut.getFilename());
@@ -122,18 +122,22 @@ public class LUTUI extends JPanel {
      * @return See above
      */
     private LookupTableItem[] createLutsArray() {
-        LookupTableItem[] lutsArray = null;
+        List<LookupTableItem> list = new ArrayList<LookupTableItem>();
+
         if (CollectionUtils.isNotEmpty(c.getAvailableLookupTables())) {
-            lutsArray = new LookupTableItem[c.getAvailableLookupTables().size() + 1];
-            lutsArray[0] = new LookupTableItem(NONE);
-            int i = 1;
             Iterator<String> it = c.getAvailableLookupTables().iterator();
             while (it.hasNext()) {
-                lutsArray[i] =  new LookupTableItem(it.next());
-                i++;
+                list.add(new LookupTableItem(it.next()));
             }
         }
-        Arrays.sort(lutsArray);
+
+        Collections.sort(list);
+
+        list.add(0, LookupTableItem.SEPARATOR);
+        list.add(0, LookupTableItem.NONE);
+
+        LookupTableItem[] lutsArray = new LookupTableItem[list.size()];
+        lutsArray = list.toArray(lutsArray);
         return lutsArray;
     }
 
@@ -151,7 +155,8 @@ public class LUTUI extends JPanel {
         int index = 0;
         if (c.getLUT() != null && lutList.getModel().getSize() > 0) {
             for (int i = 0; i < lutList.getModel().getSize(); i++) {
-                if (( (LookupTableItem)lutList.getModel().getElementAt(i) ).matchesFilename(c.getLUT())) {
+                if (((LookupTableItem) lutList.getModel().getElementAt(i))
+                        .matchesFilename(c.getLUT())) {
                     index = i;
                     break;
                 }
@@ -185,85 +190,5 @@ public class LUTUI extends JPanel {
             return;
         revert();
         repaint();
-    }
-
-    /**
-     * Item for the Lookup Table list, which shows a nicely formatted name for
-     * the lookup table, based on the file name.
-     */
-    class LookupTableItem implements Comparable<LookupTableItem> {
-
-        /** The file name **/
-        private String filename;
-
-        /** More readable name generated from the filename */
-        private String readableName;
-
-        /**
-         * Create new instance
-         * 
-         * @param filename
-         *            The lut file name
-         */
-        LookupTableItem(String filename) {
-            this.filename = filename;
-            this.readableName = generateReadableName(filename);
-        }
-
-        /**
-         * Generates a more readable name for the given lut filename by removing
-         * '*.lut' extension, underscores and using upper case at the beginning
-         * of words.
-         * 
-         * @param filename
-         *            The filename
-         * @return See above
-         */
-        private String generateReadableName(String filename) {
-            filename = filename.replace(".lut", "");
-            String[] parts = filename.replace(".lut", "").split("_");
-            StringBuilder sb = new StringBuilder();
-            for (int i = 0; i < parts.length; i++) {
-                String part = parts[i];
-                sb.append(part.substring(0, 1).toUpperCase());
-                if (part.length() > 1) {
-                    sb.append(part.substring(1));
-                }
-
-                if (i < parts.length - 1)
-                    sb.append(' ');
-            }
-
-            return sb.toString();
-        }
-
-        @Override
-        public String toString() {
-            return readableName;
-        }
-
-        /**
-         * Get the lut file name
-         * 
-         * @return See above
-         */
-        public String getFilename() {
-            return this.filename;
-        }
-
-        /**
-         * @param filename
-         *            The file name
-         * @return <code>true</code> if the given filename matches the filename
-         *         of this {@link LookupTableItem}
-         */
-        public boolean matchesFilename(String filename) {
-            return this.filename.equals(filename);
-        }
-
-        @Override
-        public int compareTo(LookupTableItem o) {
-            return this.readableName.compareTo(o.readableName);
-        }
     }
 }
